week 2/week2-rag-chromadb/Day 5/requirements.txt
openai
python-dotenv
langchain
chromadb
PyPDF2
tiktoken